requirements.txt
fasttext-wheel~=0.9.2
matplotlib~=3.9.4
networkx~=3.2.1
numpy~=1.26.4
plotly~=6.1.0
pymorphy3~=2.0.3
scikit-learn~=1.6.1
scipy~=1.13.1
spacy~=3.8.5
tqdm~=4.67.1
